move the new test to more suitable file
